@@ -14,24 +14,22 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.2.13.RELEASE</version>
-		<relativePath/> <!-- lookup parent from repository -->
+		<version>3.2.2</version>
+		<relativePath/>
 	</parent>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>1.8</java.version>
+		<java.version>17</java.version>
 
-		<shiro.version>1.13.0</shiro.version>
-		<shiro-redis.version>3.3.1</shiro-redis.version>
-		<commons-lang3.version>3.4</commons-lang3.version>
-		<commons-codec.version>1.13</commons-codec.version>
+		<commons-lang3.version>3.13.0</commons-lang3.version>
+		<commons-codec.version>1.16.0</commons-codec.version>
 		<fileupload.version>1.5</fileupload.version>
 		<mybatisplus-spring-boot-starter.version>3.5.10.1</mybatisplus-spring-boot-starter.version>
 		<mybatisplus.version>3.5.10.1</mybatisplus.version>
 		<mybatis-plus-generator.version>3.5.10.1</mybatis-plus-generator.version>
-		<fastjson.version>1.2.83</fastjson.version>
+		<fastjson.version>2.0.45</fastjson.version>
 		<jackson.version>2.16.1</jackson.version>
 		<druid.version>1.2.20</druid.version>
 		<guava.version>32.0.1-android</guava.version>
@@ -374,8 +383,8 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.11.0</version>
 				<configuration>
-					<source>1.8</source>
-					<target>1.8</target>
+					<source>17</source>
+					<target>17</target>
 					<encoding>UTF-8</encoding>
 				</configuration>
 			</plugin>
