@@ -14,7 +14,6 @@ import com.mysiteforme.admin.service.SiteService;
 import com.mysiteforme.admin.util.*;
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,8 +22,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.WebUtils;
 
-import javax.servlet.ServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.*;
 import java.nio.file.Files;
 import java.util.List;
@@ -74,7 +73,6 @@ public class TableController extends BaseController{
     /***
      * 所有数据表分页查询
      */
-    @RequiresPermissions("sys:table:list")
     @PostMapping("list")
     @ResponseBody
     public LayerData<TableVO> list(@RequestParam(value = "page",defaultValue = "1")Integer page,
@@ -99,7 +97,6 @@ public class TableController extends BaseController{
         return "admin/system/table/add";
     }
 
-    @RequiresPermissions("sys:table:add")
     @PostMapping("add")
     @ResponseBody
     @SysLog("保存数据表新增数据")
@@ -158,7 +155,6 @@ public class TableController extends BaseController{
         return "admin/system/table/edit";
     }
 
-    @RequiresPermissions("sys:table:edit")
     @PostMapping("editTable")
     @ResponseBody
     @SysLog("保存数据表编辑数据")
@@ -189,7 +185,6 @@ public class TableController extends BaseController{
         return RestResponse.success();
     }
 
-    @RequiresPermissions("sys:table:list")
     @PostMapping("fieldlist")
     @ResponseBody
     @SysLog("请求字段展示数据(分页显示)")
@@ -235,7 +230,6 @@ public class TableController extends BaseController{
         }
     }
 
-    @RequiresPermissions("sys:table:list")
     @PostMapping("showFields")
     @ResponseBody
     @SysLog("请求字段展示数据(全部显示)")
@@ -260,7 +254,6 @@ public class TableController extends BaseController{
         return tableLayerData;
     }
 
-    @RequiresPermissions("sys:table:addField")
     @PostMapping("addField")
     @ResponseBody
     @SysLog("保存单独新增字段数据")
@@ -307,7 +300,6 @@ public class TableController extends BaseController{
         return RestResponse.success();
     }
 
-    @RequiresPermissions("sys:table:editField")
     @PostMapping("editField")
     @ResponseBody
     @SysLog("保存单独编辑字段数据")
@@ -383,7 +375,6 @@ public class TableController extends BaseController{
     }
 
 
-    @RequiresPermissions("sys:table:deleteField")
     @PostMapping("deleteField")
     @ResponseBody
     @SysLog("删除字段数据")
@@ -399,7 +390,6 @@ public class TableController extends BaseController{
         return RestResponse.success();
     }
 
-    @RequiresPermissions("sys:table:deleteTable")
     @PostMapping("delete")
     @ResponseBody
     @SysLog("删除数据表数据")
@@ -414,7 +404,6 @@ public class TableController extends BaseController{
         return RestResponse.success();
     }
 
-    @RequiresPermissions("sys:table:download")
     @PostMapping("download")
     @ResponseBody
     @SysLog("下载JAVA源码")
