@@ -1,35 +1,19 @@
-package com.mysiteforme.admin.base;
-
-import com.mysiteforme.admin.entity.User;
-import com.mysiteforme.admin.realm.AuthRealm.ShiroUser;
-import com.mysiteforme.admin.service.*;
-import org.apache.shiro.SecurityUtils;
-import org.apache.shiro.subject.Subject;
-
 /**
- * 基础控制器
- * 提供通用的Controller功能
+ * @ Author: wangl
+ * @ Create Time: 2025-02-11 14:55:13
+ * @ Modified by: wangl
+ * @ Modified time: 2025-02-13 22:46:31
+ * @ Description: 基础控制器,提供通用的Controller功能
  */
+
+package com.mysiteforme.admin.base;
+import com.mysiteforme.admin.redis.RedisUtils;
+import com.mysiteforme.admin.service.*;
+import org.springframework.security.authentication.AuthenticationManager;
+
 public class BaseController {
-	
-	/**
-	 * 获取当前登录用户
-	 * @return 当前登录的用户对象,未登录返回null
-	 */
-	public User getCurrentUser() {
 
-		Subject subject = SecurityUtils.getSubject();
-		if (subject == null || !subject.isAuthenticated()) {
-			return null;
-		}
-		Object principal = subject.getPrincipal();
-		if (principal instanceof ShiroUser) {
-			ShiroUser shiroUser = (ShiroUser) principal;
-			return userService.getById(shiroUser.getId());
-		}
-		return null;
-	}
-
+	protected AuthenticationManager authenticationManager;
 	/**
 	 * 用户服务接口
 	 */
@@ -70,4 +54,5 @@ public class BaseController {
 
 	protected UploadService ossService;
 
+	protected RedisUtils redisUtils;
 }
